Split admin templates into feature-based modules
Following the same pattern used for routes and db modules, split
templates/admin.tsx into separate files under templates/admin/:
- login.tsx (login page)
- dashboard.tsx (dashboard with event list)
- events.tsx (event detail, edit, delete pages)
- attendees.tsx (attendee delete page)
- sessions.tsx (session management page)
- settings.tsx (settings page)
